1) Renamed the Memento folder to DraftStorage in the Service folder of the OutOfSchool.BusinessLogic project. 2) Renamed the appropriate Using directives.
