+++ OutOfSchool/OutOfSchool.WebApi.Tests/Services/DraftStorage/DraftStorageServiceTests.cs
@@ -7,8 +7,8 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
 using OutOfSchool.BusinessLogic.Models.Workshops.IncompletedWorkshops;
-using OutOfSchool.BusinessLogic.Services.Memento;
-using OutOfSchool.BusinessLogic.Services.Memento.Interfaces;
+using OutOfSchool.BusinessLogic.Services.DraftStorage;
+using OutOfSchool.BusinessLogic.Services.DraftStorage.Interfaces;
 using OutOfSchool.Redis;
 
 namespace OutOfSchool.WebApi.Tests.Services.DraftStorage;
