+++ OutOfSchool/OutOfSchool.WebApi/Startup.cs
@@ -16,9 +16,9 @@ using OutOfSchool.BackgroundJobs.Config;
 using OutOfSchool.BackgroundJobs.Extensions.Startup;
 using OutOfSchool.BusinessLogic.Services.AverageRatings;
 using OutOfSchool.BusinessLogic.Services.Communication.ICommunication;
-using OutOfSchool.BusinessLogic.Services.Memento;
-using OutOfSchool.BusinessLogic.Services.Memento.Interfaces;
-using OutOfSchool.BusinessLogic.Services.Memento.JSONConverter;
+using OutOfSchool.BusinessLogic.Services.DraftStorage;
+using OutOfSchool.BusinessLogic.Services.DraftStorage.Interfaces;
+using OutOfSchool.BusinessLogic.Services.DraftStorage.JSONConverter;
 using OutOfSchool.BusinessLogic.Services.ProviderServices;
 using OutOfSchool.BusinessLogic.Services.Strategies.Interfaces;
 using OutOfSchool.BusinessLogic.Services.Strategies.WorkshopStrategies;
